remove hklpy2 labels from diffractometer
